Updated backend

+++ graphql/resolvers/booking.js
@@ -1,43 +1,44 @@
-const Booking= require('../../models/booking');
+const Event = require('../../models/event');
+const Booking = require('../../models/booking');
 const { transformBooking, transformEvent } = require('./merge');
-const Event= require('../../models/event');
 
-
-
-
-
-module.exports= {
-bookings: async ()=>{
-    try{
-        const bookings= await Booking.find();
-        return bookings.map(booking =>{
-            return transformBooking(booking);
-    });
-    }catch(err){
-        throw err;
+module.exports = {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
     }
-},
-    bookEvent: async (args) =>{
-        const fetchedEvent= await Event.findOne({_id: args.eventId})
-       const booking= new Booking({
-        user: '5dee1f0484952c14c4a2d0c5',
-        event: fetchedEvent
+    try {
+      const bookings = await Booking.find();
+      return bookings.map(booking => {
+        return transformBooking(booking);
+      });
+    } catch (err) {
+      throw err;
+    }
+  },
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    const fetchedEvent = await Event.findOne({ _id: args.eventId });
+    const booking = new Booking({
+      user: req.userId,
+      event: fetchedEvent
     });
-    const result= await booking.save();
+    const result = await booking.save();
     return transformBooking(result);
-    },
-    cancelBooking : async (args) =>{
-        try{
-            const booking= await Booking.findById(args.bookingId).populate('event');
-            const event= { 
-                ...booking.event._doc, 
-                _id: booking.event.id, 
-                creator: user.bind(this, booking.event._doc.creator)
-            }
-            await Booking.deleteOne({_id: args.bookingId})
-            return event;
-        }catch(err){
-            throw err;
-        }
+  },
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
     }
-}
+    try {
+      const booking = await Booking.findById(args.bookingId).populate('event');
+      const event = transformEvent(booking.event);
+      await Booking.deleteOne({ _id: args.bookingId });
+      return event;
+    } catch (err) {
+      throw err;
+    }
+  }
+};
